requirements.txt
# Airflow constraint-compatible versions
apache-airflow==2.9.3
apache-airflow-providers-common-io>=1.3.0

# Processing
duckdb==0.10.3
pyarrow==16.1.0
pandas==2.2.2

# Data Quality
great-expectations==0.18.19

# Utilities
python-dotenv==1.0.1

# Testing
pytest>=8.2
pytest-cov>=5.0
